UI completed. All UI elements are minimal now and all expected actions act as intended.
