+++ src/App.css
@@ -1,66 +1,7 @@
-.app-container {
-  min-height: 100vh;
-  background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
-}
-
-.app-header {
-  text-align: center;
-  padding: 2rem 1rem 1rem;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.app-header h1 {
-  font-size: 2.5rem;
-  margin-bottom: 0.5rem;
-  background: linear-gradient(45deg, #646cff, #61dafb);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-}
-
-.app-header p {
-  color: #888;
-  font-size: 1.1rem;
-  margin: 0;
-  opacity: 0.8;
-}
-
-.app-main {
-  display: flex;
-  justify-content: center;
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 2rem 1rem;
-}
-
-.voice-section {
-  background: rgba(255, 255, 255, 0.03);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 16px;
-  padding: 2rem;
-  backdrop-filter: blur(10px);
-  width: 100%;
-  max-width: 600px;
-}
 
 body {
   background: #1a1a1a;
   color: #ffffff;
   margin: 0;
   min-height: 100vh;
-}
-
-@media (max-width: 768px) {
-  .app-header h1 {
-    font-size: 2rem;
-  }
-
-  .app-main {
-    padding: 1rem;
-  }
-
-  .voice-section {
-    padding: 1.5rem;
-  }
 }
