+++ src/App.jsx
@@ -1,10 +1,8 @@
 
-import { useState } from 'react';
-import TTSPanel from "./components/TTSPanel";
-import ChatPanel from './components/ChatPanel.jsx';
 
+import ChatPanel from './components/ChatPanel.jsx';
 import './App.css'
-import { VoiceAssistant } from './components/VoiceAssistant.jsx'
+
 
 function App() {
 
@@ -12,18 +10,7 @@ function App() {
 
   return (
     <>
-      {/* <div className="app-container">
-        <header className="app-header">
-          <h1>Jarvis - Voice Assistant</h1>
-          <p>Speak to AI, get intelligent responses</p>
-        </header>
-
-        <main className="app-main">
-          <section className="voice-section">
-            <VoiceAssistant />
-          </section>
-        </main>
-      </div> */}
+  
 
       <ChatPanel />
