+++ src/components/ChatPanel.css
@@ -117,6 +117,43 @@
   }
 }
 
+.avatar-orb.listening {
+  animation: orb-pulse 1.5s ease-in-out infinite;
+  box-shadow:
+    0 0 40px rgba(255, 107, 53, 0.6),
+    0 0 80px rgba(196, 69, 105, 0.4),
+    inset 0 0 40px rgba(247, 147, 30, 0.3);
+}
+
+@keyframes orb-pulse-speaking {
+
+  0%,
+  100% {
+    transform: scale(1);
+    box-shadow:
+      0 0 30px rgba(0, 230, 230, 0.8),
+      0 0 60px rgba(26, 117, 255, 0.5),
+      inset 0 0 30px rgba(212, 1, 202, 0.4);
+  }
+
+  50% {
+    transform: scale(1.1);
+    box-shadow:
+      0 0 40px rgba(0, 230, 230, 1),
+      0 0 80px rgba(26, 117, 255, 0.7),
+      inset 0 0 40px rgba(212, 1, 202, 0.6);
+  }
+}
+
+.avatar-orb.speaking {
+  background: linear-gradient(135deg, #00e6e6 0%, #1affd5 40%, #1a75ff 80%, #d401ca 100%);
+  animation: orb-pulse-speaking 1.5s ease-in-out infinite;
+  box-shadow:
+    0 0 40px rgba(0, 230, 230, 0.8),
+    0 0 80px rgba(26, 117, 255, 0.5),
+    inset 0 0 50px rgba(212, 1, 202, 0.4);
+}
+
 .chat-button {
   background: transparent;
   color: #fe8019;
@@ -127,26 +164,28 @@
   font-weight: 700;
   cursor: pointer;
   transition: all 0.3s ease;
-  box-shadow: 0 0px 10px rgba(255, 167, 100, 0.3);
-  text-transform: uppercase;
+  /* box-shadow: 0 0px 10px rgba(255, 167, 100, 0.3); */
   letter-spacing: 1px;
 }
 
-.chat-button:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 8px 25px rgba(255, 185, 100, 0.4);
+.chat-button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  transform: none;
+  box-shadow: none;
 }
 
+.chat-button.listening {
+  background: linear-gradient(45deg, #ff6b6b, #ff8e8e);
+  color: white;
+}
+
+
 .chat-history-section {
-  background: rgba(42, 42, 62, 0.3);
-  border-radius: 15px;
-  padding: 20px;
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  box-shadow: 0 8px 10px rgba(0, 0, 0, 0.3);
   flex: 2;
   min-width: 400px;
   max-width: 800px;
-  width: 100%;
+
 
 }
 
@@ -155,7 +194,12 @@
   overflow-y: auto;
   margin-bottom: 20px;
   padding-right: 15px;
-  
+  background: rgba(42, 42, 62, 0.3);
+  border-radius: 15px;
+  padding: 20px;
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  box-shadow: 0 8px 10px rgba(0, 0, 0, 0.3);
+
 }
 
 .chat-messages::-webkit-scrollbar {
@@ -173,18 +217,17 @@
 }
 
 .message {
-  margin-bottom: 15px;
   padding: 15px;
   border-radius: 15px;
 }
 
 
-.message.you {
-  text-align: right;
-  background: rgba(63, 81, 181, 0.1);
+.message-you {
+  text-align: left;
+  /* background: rgba(63, 81, 181, 0.1); */
 }
 
-.message.jarvis {
+.message-jarvis {
   text-align: left;
 }
 
@@ -193,8 +236,12 @@
   line-height: 1.5;
 }
 
-.message-content strong {
+.message-you strong {
   color: #ffaf64;
+}
+
+.message-jarvis strong {
+  color: #d401ca;
 }
 
 .message-time {
@@ -212,7 +259,7 @@
 }
 
 .clear-chat-button {
- 
+
   background: transparent;
   color: rgb(255, 0, 0);
   border: none;
@@ -226,7 +273,12 @@
 
 .clear-chat-button:hover {
   transform: translateY(-2px);
-  
+}
+
+.clear-chat-button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  transform: none;
 }
 
 .chat-footer {
