--- src/components/VoiceAssistant.css
@@ -1,235 +1,0 @@
-.voice-assistant {
-    width: 100%;
-    text-align: center;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-}
-
-.voice-assistant header h1 {
-    font-size: 1.5rem;
-    margin-bottom: 0.5rem;
-    background: linear-gradient(45deg, #646cff, #61dafb);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-}
-
-.status {
-    font-size: 0.9rem;
-    color: #888;
-    margin-bottom: 1rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.5rem;
-    flex-wrap: wrap;
-}
-
-.auto-speak-toggle {
-    display: flex;
-    align-items: center;
-    gap: 0.25rem;
-    cursor: pointer;
-    font-size: 0.85rem;
-}
-
-.auto-speak-toggle input {
-    margin: 0;
-}
-
-main {
-    margin: 2rem 0;
-}
-
-.controls {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    align-items: center;
-    margin: 1.5rem 0;
-}
-
-.voice-btn {
-    background: linear-gradient(45deg, #646cff, #61dafb);
-    color: white;
-    border: none;
-    padding: 1rem 2rem;
-    font-size: 1.1rem;
-    border-radius: 50px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 15px rgba(100, 108, 255, 0.3);
-    min-width: 200px;
-    font-weight: 500;
-}
-
-.voice-btn.primary {
-    font-size: 1.2rem;
-    padding: 1.2rem 2.5rem;
-    min-width: 250px;
-}
-
-.voice-btn.secondary {
-    background: linear-gradient(45deg, #555, #777);
-    font-size: 1rem;
-    padding: 0.8rem 1.8rem;
-    min-width: 180px;
-    box-shadow: 0 3px 10px rgba(0, 0, 0, 0.3);
-}
-
-.voice-btn.clear {
-    background: linear-gradient(45deg, #ff6b6b, #ff8e8e);
-}
-
-.voice-btn:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 6px 20px rgba(100, 108, 255, 0.4);
-}
-
-.voice-btn:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    transform: none;
-}
-
-.voice-btn.listening {
-    background: linear-gradient(45deg, #ff6b6b, #ff8e8e);
-    animation: pulse 1.5s infinite;
-}
-
-.voice-btn.disabled {
-    background: #555;
-    cursor: not-allowed;
-}
-
-.voice-indicator-container {
-    margin: 2rem 0;
-    display: flex;
-    justify-content: center;
-}
-
-@media (max-width: 768px) {
-    .controls {
-        gap: 0.8rem;
-    }
-    
-    .voice-btn.primary {
-        font-size: 1.1rem;
-        padding: 1rem 2rem;
-        min-width: 220px;
-    }
-    
-    .voice-btn.secondary {
-        font-size: 0.9rem;
-        padding: 0.7rem 1.5rem;
-        min-width: 160px;
-    }
-}
-
-@keyframes pulse {
-    0%,
-    100% {
-        opacity: 1;
-    }
-
-    50% {
-        opacity: 0.7;
-    }
-}
-
-.listening {
-    color: #ff6b6b;
-    font-size: 1.1rem;
-    margin: 1rem 0;
-    animation: pulse 1s infinite;
-}
-
-.speaking {
-    color: #4caf50;
-    font-size: 1.1rem;
-    margin: 1rem 0;
-    animation: pulse 1s infinite;
-}
-
-.error {
-    color: #ff6b6b;
-    background: rgba(255, 107, 107, 0.1);
-    padding: 1rem;
-    border-radius: 8px;
-    margin: 1rem 0;
-    font-size: 0.9rem;
-}
-
-.result {
-    background: rgba(255, 255, 255, 0.05);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    border-radius: 12px;
-    padding: 1rem;
-    margin: 1rem 0;
-    text-align: left;
-}
-
-.result strong {
-    color: #646cff;
-}
-
-.history {
-    background: rgba(255, 255, 255, 0.03);
-    border: 1px solid rgba(255, 255, 255, 0.08);
-    border-radius: 12px;
-    padding: 1rem;
-    margin: 2rem 0;
-    text-align: left;
-}
-
-.history h3 {
-    color: #646cff;
-    margin-bottom: 1rem;
-    font-size: 1rem;
-}
-
-.history-item {
-    background: rgba(255, 255, 255, 0.02);
-    padding: 0.8rem;
-    border-radius: 6px;
-    margin-bottom: 0.8rem;
-    border-left: 3px solid #646cff;
-}
-
-.history-item small {
-    color: #888;
-    font-size: 0.8rem;
-}
-
-.history {
-  background: rgba(255, 255, 255, 0.03);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 12px;
-  padding: 1rem;
-  margin: 2rem 0;
-  text-align: left;
-}
-
-.history h3 {
-  color: #646cff;
-  margin-bottom: 1rem;
-  font-size: 1rem;
-}
-
-.history-item {
-  background: rgba(255, 255, 255, 0.02);
-  padding: 0.8rem;
-  border-radius: 6px;
-  margin-bottom: 0.8rem;
-  border-left: 3px solid #646cff;
-}
-
-.history-item small {
-  color: #888;
-  font-size: 0.8rem;
-}
-
-.voice-indicator-container {
-    margin: 2rem 0;
-    display: flex;
-    justify-content: center;
-}
