--- src/components/VoiceIndicator.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import './VoiceIndicator.css';
-
-const VoiceIndicator = ({ isPlaying = false, size = 60 }) => {
-  return (
-    <div className="voice-indicator">
-      <div 
-        className={`voice-circle ${isPlaying ? 'speaking' : 'idle'}`}
-        style={{ width: size, height: size }}
-      >
-        <div className="inner-circle"></div>
-      </div>
-    </div>
-  );
-};
-
-export default VoiceIndicator;
